requirements.txt
fastapi==0.115.11
uvicorn==0.30.6
httpx==0.27.2
pyyaml==6.0.2
jinja2==3.1.5
python-multipart==0.0.20
